Tried making clock / timer, read description
Looked it up on google and found some code that could work, but "sleep_for" freezes the whole program before continueing.
This isn't what I want, because I want to have the game never freeze, just a timer on the code before it continues. Kinda something simular to Unity's WaitForSeconds